+++ Chapter.Net.WPF.Converters.Tests/ConcatenateStringConverter/ConcatenateStringConverterTests.cs
@@ -11,14 +11,19 @@ using NUnit.Framework;
 
 namespace Chapter.Net.WPF.Converters.Tests;
 
-public class ConcatenateStringConverterTests : MultiConverterTester<ConcatenateStringConverter>
+public class ConcatenateStringConverterTests : MultiValueConverterTester<ConcatenateStringConverter>
 {
-    [TestCase("a..b..c", "..", "a", "b", "c")]
-    [TestCase("a..1..4", "..", "a", 1, 4)]
-    [TestCase("a.b.c", ".", "a", null, "b", "c", null)]
-    [TestCase("", "/", null, null)]
-    public void Convert_Called_Concatenates(string expectation, string separator, params object[] values)
+    [TestCase(false, "a..b..c", "..", "a", "b", "c")]
+    [TestCase(false, "a..1..4", "..", "a", 1, 4)]
+    [TestCase(false, "a.b.c", ".", "a", null, "b", "c", null)]
+    [TestCase(false, "", "/", null, null)]
+    [TestCase(true, "a..b..c", "..", "a", "b", "c")]
+    [TestCase(true, "a..1..4", "..", "a", 1, 4)]
+    [TestCase(true, "a..b.c.", ".", "a", null, "b", "c", null)]
+    [TestCase(true, "/", "/", null, null)]
+    public void Convert_Called_Concatenates(bool acceptNullParts, string expectation, string separator, params object[] values)
     {
+        _target.AcceptNullParts = acceptNullParts;
         _target.Separator = separator;
 
         Convert(values, expectation);